requirements.txt
flask
python-dotenv
psycopg
attrs==21.4.0
beautifulsoup4==4.10.0
bs4==0.0.1
certifi==2021.10.8
charset-normalizer==2.0.12
idna==3.3
lxml==4.8.0
numpy==1.22.3
pandas==1.4.1
price-parser==0.3.4
python-dateutil==2.8.2
pytz==2021.3
requests==2.27.1
six==1.16.0
soupsieve==2.3.1
urllib3==1.26.9